requirements.txt
svgwrite >= 1.2.1
pyyaml >= 3.13
